+++ RossQoL/src/RossQoL.Game/Framework/FeatureRegistry.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using RossQoL.Game.Crafting;
 using RossQoL.Game.Portals;
 using RossQoL.Game.Startup;
 
@@ -12,6 +13,8 @@ namespace RossQoL.Game.Framework
     {
         public static IReadOnlyList<Category> Create() => new[]
         {
+            new Category("Crafting", "All crafting station tweaks.",
+                new CraftingSearchFeature()),
             new Category("Portals", "All portal tweaks.",
                 new TamesFollowFeature()),
             new Category("Startup", "All startup and main menu tweaks.",